requirements.txt
SQLAlchemy>=2.0.41
pydantic>=2.11.4
numpy>=2.2.5
pandas>=2.2.3
beautifulsoup4>=4.13.4
selenium>=4.32.0
dotenv>=0.9.9
python-dotenv>=1.1.0
pyyaml>=6.0.2
fastapi>=0.115.12
django>=5.2.1